+++ lib/govuk_notify_rails/mail_ext.rb
@@ -1,6 +1,7 @@
 module Mail
   class Message
     attr_accessor :govuk_notify_template
+    attr_accessor :govuk_notify_reference
     attr_accessor :govuk_notify_personalisation
   end
 end