+++ spec/dummy/app/mailers/notify_mailer.rb
@@ -1,6 +1,7 @@
 class NotifyMailer < GovukNotifyRails::Mailer
   def test_email(user)
     set_template('9661d08a-486d-4c67-865e-ad976f17871d')
+    set_reference('my_reference')
     set_personalisation(
       full_name: user.name
     )