+++ spec/dummy/config/initializers/govuk_notify_rails.rb
@@ -1,3 +1,2 @@
 ActionMailer::Base.add_delivery_method :govuk_notify, GovukNotifyRails::Delivery,
-                                       service_id: 'test-service-id',
-                                       secret_key: 'test-secret-key'
+                                       api_key: 'test-api-key'